+++ fcft/fcft.h
@@ -5,29 +5,54 @@
 
 #include <pixman.h>
 
-struct glyph {
+/*
+ * Defines the subpixel order to use.
+ *
+ * Note that this is *ignored* if antialiasing has been disabled.
+ */
+enum fcft_subpixel {
+    FCFT_SUBPIXEL_DEFAULT,          /* Use subpixel order from FontConfig */
+    FCFT_SUBPIXEL_NONE,             /* Disable subpixel antialiasing (use grayscale antialiasing) */
+    FCFT_SUBPIXEL_HORIZONTAL_RGB,
+    FCFT_SUBPIXEL_HORIZONTAL_BGR,
+    FCFT_SUBPIXEL_VERTICAL_RGB,
+    FCFT_SUBPIXEL_VERTICAL_BGR,
+};
+
+struct fcft_glyph {
     wchar_t wc;
     int cols;              /* wcwidth(wc) */
 
     pixman_image_t *pix;
+
     int x;
     int y;
-    int x_advance;
     int width;
     int height;
 
-    /* Internal */
-    bool subpixel_antialias;
-    bool valid;
+    struct {
+        int x;
+        int y;
+    } advance;
 };
 
-struct font {
+struct fcft_font {
     /* font extents */
     int height;
     int descent;
     int ascent;
-    int max_x_advance;    /* Width of font's widest glyph */
-    int space_x_advance;  /* Width of space (0x20), if available, -1 otherwise */
+
+    /* Width/height of font's widest glyph */
+    struct {
+        int x;
+        int y;
+    } max_advance;
+
+    /* Width/height of space (0x20), if available, -1 otherwise */
+    struct {
+        int x;
+        int y;
+    } space_advance;
 
     struct {
         int position;
@@ -42,17 +67,19 @@ struct font {
 
 /* First entry is the main/primary font, the remaining (if any) are
  * custom fallback fonts */
-struct font *font_from_name(
+struct fcft_font *fcft_from_name(
     size_t count, const char *names[static count], const char *attributes);
-struct font *font_clone(const struct font *font);
-void font_destroy(struct font *font);
+struct fcft_font *fcft_clone(const struct fcft_font *font);
+void fcft_destroy(struct fcft_font *font);
 
 /* Returns a *new* font instance */
-struct font *font_size_adjust(const struct font *font, double amount);
+struct fcft_font *fcft_size_adjust(const struct fcft_font *font, double amount);
 
-const struct glyph *font_glyph_for_wc(
-    struct font *font, wchar_t wc, bool subpixel_antialias);
+/* Rasterize 'wc' using 'font'. Use the defined subpixel mode *if*
+ * antialiasing is enabled for this font */
+const struct fcft_glyph *fcft_glyph_rasterize(
+    struct fcft_font *font, wchar_t wc, enum fcft_subpixel subpixel);
 
-bool font_kerning(
-    struct font *font, wchar_t left, wchar_t right,
+bool fcft_kerning(
+    struct fcft_font *font, wchar_t left, wchar_t right,
     long *restrict x, long *restrict y);