+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 * [Unreleased](#unreleased)
+* [2.0.0](#2-0-0)
 * [1.1.7](#1-1-7)
 
 ## Unreleased
@@ -10,6 +11,33 @@
 ### Removed
 ### Fixed
 ### Security
+
+
+## 2.0.0
+
+### Changed
+
+* API: `font_` prefix changed to `fcft_`.
+* API: renamed `struct font` to `struct fcft_font`.
+* API: renamed `struct glyph` to `struct fcft_glyph`.
+* API: internal members of `struct fcft_glyph` removed.
+* API: renamed `enum subpixel_order` to `enum fcft_subpixel`, and
+  `ORDER` was removed from the enum values.
+* API: renamed `fcft_glyph.x_advance` to `fcft_glyph.advance.x`, and
+  added `fcft_glyph.advance.y`
+* API: renamed `fcft_font.max_x_advance` to `fcft_font.max_advance.x`
+  and added `fcft_font.max_advance.y`.
+* API: renamed `fcft_font.space_x_advance` to
+  `fcft_font.space_advance.x` and added `fcft_font.space_advance.y`.
+* API: renamed `fcft_glyph_for_wc()` to `fcft_glyph_rasterize()`.
+* Require meson >= 0.54.
+* Use `meson.override_dependency()`.
+
+
+### Fixed
+
+* `fcft_kerning()` did not scale the returned kerning distances with
+  the font's pixel size fixup multiplier.
 
 
 ## 1.1.7